refactor: move lz4 and lzf revobject serialization decorators to their own modules
Most probably a first step towards deprecating and removing
them, moving them off as direct dependencies of core specially
because they carry over native libraries that might better
be avoided depending on the library components needed on each
application that uses geogig as a library.
