+++ src/core/pom.xml
@@ -70,17 +70,6 @@
       <artifactId>guice</artifactId>
       <classifier>no_aop</classifier>
     </dependency>
-
-    <dependency>
-      <groupId>com.ning</groupId>
-      <artifactId>compress-lzf</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>net.jpountz.lz4</groupId>
-      <artifactId>lz4</artifactId>
-    </dependency>
-
     <!-- Test scope dependencies -->
     <dependency>
       <groupId>junit</groupId>
