+++ src/storage/formats/lz4/src/test/java/org/locationtech/geogig/storage/format/lz4/RevObjectSerializerLZ4Test.java
@@ -7,15 +7,16 @@
  * Contributors:
  * Gabriel Roldan (Boundless) - initial implementation
  */
-package org.locationtech.geogig.storage.datastream;
+package org.locationtech.geogig.storage.format.lz4;
 
 import org.locationtech.geogig.storage.RevObjectSerializer;
+import org.locationtech.geogig.storage.datastream.DataStreamRevObjectSerializerV2_1;
 import org.locationtech.geogig.storage.impl.RevObjectSerializerConformanceTest;
 
-public class RevObjectSerializerLZFTest extends RevObjectSerializerConformanceTest {
+public class RevObjectSerializerLZ4Test extends RevObjectSerializerConformanceTest {
 
     protected @Override RevObjectSerializer newObjectSerializer() {
-        return new RevObjectSerializerLZF(DataStreamRevObjectSerializerV2_1.INSTANCE);
+        return new RevObjectSerializerLZ4(DataStreamRevObjectSerializerV2_1.INSTANCE);
     }
 
 }
