+++ src/storage/postgres/pom.xml
@@ -49,6 +49,12 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <!-- for legacy purposes, may be removed in the future -->
+      <groupId>org.locationtech.geogig</groupId>
+      <artifactId>geogig-serialization-lzf</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>org.postgresql</groupId>
       <artifactId>postgresql</artifactId>
     </dependency>
