+++ src/storage/rocksdb/pom.xml
@@ -50,6 +50,12 @@
       <version>${project.version}</version>
     </dependency>
     <dependency>
+      <!-- for legacy purposes, may be removed in the future -->
+      <groupId>org.locationtech.geogig</groupId>
+      <artifactId>geogig-serialization-lzf</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <!-- Google Common Libraries. Featuring com.google.collect collection classes -->
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
